[FEATURE] FragmentStore now allows to specify the StringSet specs of the ReadNameStore
git-svn-id: http://svn.seqan.de/seqan/trunk@12511 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/core/apps/splazers/outputFormat.h
@@ -1599,7 +1599,7 @@ void dumpMatches(
 				TMatch& mR = *it;
 				unsigned readLenL = mL.mScore;
 				unsigned readLenR = mR.mScore;
-				SEQAN_ASSERT_LEQ(mL.mScore+mR.mScore,readLen);
+				SEQAN_ASSERT_LEQ(mL.mScore+mR.mScore,(int)readLen);
 				String<Dna5Q> readInfL = infix(reads[currReadNo],0,readLenL);
 				String<Dna5Q> readInfR = infix(reads[currReadNo],length(reads[currReadNo])-readLenR,length(reads[currReadNo]));
 				int offsetL = 0;
